+++ e2e-tests/commands/insurance/complete-and-submit-all-insurance-eligibility-answers.js
@@ -19,8 +19,8 @@ const completeAndSubmitAllInsuranceEligibilityAnswers = ({ companyNumber = COMPA
   cy.completeCoverPeriodForm({});
   cy.completeUkGoodsAndServicesForm();
   cy.completeEndBuyerForm();
-  cy.completePartyToConsortiumForm();
-  cy.completeMemberOfAGroupForm();
+  cy.completePartyToConsortiumForm({});
+  cy.completeMemberOfAGroupForm({});
 };
 
 export default completeAndSubmitAllInsuranceEligibilityAnswers;